+++ EmbeddedSocialClient/sdk/src/main/java/com/microsoft/embeddedsocial/ui/activity/PopularActivity.java
@@ -8,7 +8,7 @@ package com.microsoft.embeddedsocial.ui.activity;
 import com.microsoft.embeddedsocial.ui.activity.base.BaseActivity;
 import com.microsoft.embeddedsocial.ui.fragment.FeedViewMenuFragment;
 import com.microsoft.embeddedsocial.sdk.R;
-import com.microsoft.embeddedsocial.ui.fragment.PopularFeedFragmentTabs;
+import com.microsoft.embeddedsocial.ui.fragment.PopularFeedFragment;
 
 /**
  * Activity showing popular feeds.
@@ -22,7 +22,7 @@ public class PopularActivity extends BaseActivity {
 
 	@Override
 	protected void setupFragments() {
-		setActivityContent(new PopularFeedFragmentTabs());
+		setActivityContent(new PopularFeedFragment());
 		super.setupFragments();
 		getSupportFragmentManager().beginTransaction().add(new FeedViewMenuFragment(), FeedViewMenuFragment.TAG).commit();
 	}
