+++ EmbeddedSocialClient/sdk/src/main/java/com/microsoft/embeddedsocial/ui/fragment/RecentActivityFragment.java
@@ -13,7 +13,10 @@ import com.microsoft.embeddedsocial.ui.util.SimplePagerAdapter;
 
 import android.support.v4.view.PagerAdapter;
 
-public class RecentActivityFragmentTabs extends BaseTabsFragment {
+/**
+ * Fragment showing recent activity feeds displayed as tabbed views
+ */
+public class RecentActivityFragment extends BaseTabsFragment {
     @Override
     protected PagerAdapter createPagerAdapter() {
         SimplePagerAdapter.Page userFeed = new SimplePagerAdapter.Page(R.string.es_activity_feed_user, UserActivityFeedFragment::new);
